+++ src/rmp/abstract_connector.py
